updates to baopoco
upped the number of quantization bits to 18bits (full precision).
that is from 8bits/sample(4 real, 4 imag) to 36bits/sample(18 real,
18 imag). This required changes to the quant block and the xengine
blocks(specifically slices to input of xengine. the complex multipliers
in cross mult and changed bits in vacc from 32_6 to 32_30.)
Also added trigger for the adc_capture. must go from 0 to 1.