huge error in baopoco design...
specifically error in equaliser block.
We were comparing a 36_20 signed number
with a signed 4_3 (7/8, -7/8) number to generate
errors. This lacked bits on the output.
Basically, everything was getting clipped. the rms values
out of this block were ~7 regardless of what the input was (digital noise,
undriven input, or analog noise).