renamed outputs of bitbasher in the transpose block ....
that splits the incoming channel data into even and odd channels.
Previous design had these switched up.

Also fixed up slices in equaliser blocks. Channel 0 coeff was going to
channel 1, and channel 1 coeff was going to channel 0... coeff was
switched for each pair of adjacent channels. This is fixed in this commit.